modal and css tomfoolery

+++ components/Header.tsx
@@ -11,7 +11,6 @@ type HeaderProps = {
   canNext: boolean
 }
 
-// TODO: add jump to "today" button if not on current day
 const Header = ({ date, onPrev, onNext, canPrev, canNext }: HeaderProps) => {
   return (
     <div className="">
@@ -20,7 +19,7 @@ const Header = ({ date, onPrev, onNext, canPrev, canNext }: HeaderProps) => {
       <div className="flex">
         <div className={'p-4 rounded-full ' + (canPrev ? 'text-black' : 'text-slate-300')} onClick={canPrev ? onPrev : undefined}><ChevronLeft style={{ width: 36, height: 36 }} /></div>
         <div className="grow text-2xl font-semibold text-center leading-[calc(36px+2rem)]">
-          {date.toLocaleDateString('de-DE', { weekday: 'long' })} - {formatDate(date)}
+          {date.toLocaleDateString('de-DE', { weekday: 'short' })} - {formatDate(date)}
         </div>
         <div className={'p-4 rounded-full ' + (canNext ? 'text-black' : 'text-slate-300')} onClick={canNext ? onNext : undefined}><ChevronRight style={{ width: 36, height: 36 }} /></div>
       </div>
